Unit test for interface
