+++ src/backend/simulator/tests/architecture/__init__.py
@@ -9,3 +9,4 @@ sys.path.insert(0, '../../../')
 from simulator.tests.architecture.limits_t import Limits_t
 from simulator.tests.architecture.architecture_t import Architecture_t
 from simulator.tests.architecture.package_manager_t import PackageManager_t
+from simulator.tests.architecture.interface_t import Interface_t
